update code according to comments
