--- src/nni_manager/training_service/kubernetes/frameworkcontroller/frameworkcontrollerData.ts
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation
- * All rights reserved.
- *
- * MIT License
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
- * to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
- * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-'use strict';
-
-
-export const FrameworkControllerScriptFormat =
-`#!/bin/bash
-export NNI_PLATFORM={0}
-export NNI_SYS_DIR=$PWD/nni/{1}
-export NNI_OUTPUT_DIR={2}
-export MULTI_PHASE=false
-export NNI_TRIAL_JOB_ID={3}
-export NNI_EXP_ID={4}
-export NNI_CODE_DIR={5}
-export NNI_TRIAL_SEQ_ID={6}
-{7}
-mkdir -p $NNI_SYS_DIR
-mkdir -p $NNI_OUTPUT_DIR
-cp -rT $NNI_CODE_DIR $NNI_SYS_DIR
-cd $NNI_SYS_DIR
-sh install_nni.sh
-{8}
-. /mnt/frameworkbarrier/injector.sh
-eval "python3 -m nni_trial_tool.trial_keeper --trial_command '{9}' --nnimanager_ip {10} --nnimanager_port {11} `
-+ `1>$NNI_OUTPUT_DIR/trialkeeper_stdout 2>$NNI_OUTPUT_DIR/trialkeeper_stderr"`
