+++ src/nni_manager/training_service/kubernetes/frameworkcontroller/frameworkcontrollerTrainingService.ts
@@ -41,21 +41,19 @@ import { FrameworkControllerTrialConfig } from './frameworkcontrollerConfig';
 import { FrameworkControllerJobRestServer } from './frameworkcontrollerJobRestServer';
 import { FrameworkControllerClient } from './frameworkcontrollerApiClient';
 import { FrameworkControllerJobInfoCollector } from './frameworkcontrollerJobInfoCollector';
-import { FrameworkControllerScriptFormat } from './frameworkcontrollerData';
-import { String } from 'typescript-string-operations';
 
 /**
  * Training Service implementation for frameworkcontroller
  */
 @component.Singleton
 class FrameworkControllerTrainingService extends KubernetesTrainingService implements KubernetesTrainingService {
-    private frameworkcontrollerTrialConfig?: FrameworkControllerTrialConfig;
-    private frameworkcontrollerJobInfoCollector: FrameworkControllerJobInfoCollector;
-    private frameworkcontrollerContainerPortMap = new Map<string, number>();
+    private fcTrialConfig?: FrameworkControllerTrialConfig; // frameworkcontroller trial configuration
+    private fcJobInfoCollector: FrameworkControllerJobInfoCollector; // frameworkcontroller job info collector
+    private fcContainerPortMap = new Map<string, number>(); // store frameworkcontroller container port
 
     constructor() {
         super();
-        this.frameworkcontrollerJobInfoCollector = new FrameworkControllerJobInfoCollector(this.trialJobsMap);
+        this.fcJobInfoCollector = new FrameworkControllerJobInfoCollector(this.trialJobsMap);
         this.experimentId = getExperimentId();      
         this.nextTrialSequenceId = -1;
     }
@@ -70,7 +68,7 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
         while (!this.stopping) {
             // collect metrics for frameworkcontroller jobs by interacting with Kubernetes API server  
             await delay(3000);
-            await this.frameworkcontrollerJobInfoCollector.retrieveTrialStatus(this.kubernetesCRDClient);
+            await this.fcJobInfoCollector.retrieveTrialStatus(this.kubernetesCRDClient);
         }
     }
 
@@ -161,64 +159,40 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
         return Promise.resolve(trialJobOutputUrl);
     }
     
-    private generatePortScript(): string {
+    /**
+     * generate trial's command for frameworkcontroller
+     * expose port and execute injector.sh before user's command
+     * @param command 
+     */
+    private generateCommandScript(command: string): string {
         let portScript = '';
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
-
-        for(let index in this.frameworkcontrollerTrialConfig.taskRoles) {
+        for(let index in this.fcTrialConfig.taskRoles) {
             portScript += 
-            `${this.frameworkcontrollerTrialConfig.taskRoles[index].name}_port=${this.frameworkcontrollerContainerPortMap.get(this.frameworkcontrollerTrialConfig.taskRoles[index].name)} `;
+            `${this.fcTrialConfig.taskRoles[index].name}_port=${this.fcContainerPortMap.get(this.fcTrialConfig.taskRoles[index].name)} `;
         }
-        return portScript;
-    }
-
-    private generateRunScript(platform: string, trialJobId: string, trialWorkingFolder: string, 
-        command: string, trialSequenceId: string, roleName: string, gpuNum: number): string {
-        let portScript: string = this.generatePortScript();
-        let nvidia_script: string = '';
-        // Nvidia devcie plugin for K8S has a known issue that requesting zero GPUs allocates all GPUs
-        // Refer https://github.com/NVIDIA/k8s-device-plugin/issues/61
-        // So we have to explicitly set CUDA_VISIBLE_DEVICES to empty if user sets gpuNum to 0 in NNI config file
-        if(gpuNum === 0) {
-            nvidia_script = `export CUDA_VISIBLE_DEVICES='0'`;
-        }
-        const nniManagerIp = this.nniManagerIpConfig?this.nniManagerIpConfig.nniManagerIp:getIPV4Address();
-        const runScript: string = String.Format(
-            FrameworkControllerScriptFormat,
-            platform,
-            trialJobId,
-            path.join(trialWorkingFolder, 'output', `${roleName}_output`),
-            trialJobId,
-            getExperimentId(),
-            trialWorkingFolder,
-            trialSequenceId,
-            nvidia_script,
-            portScript,
-            command,
-            nniManagerIp,
-            this.kubernetesRestServerPort
-        );
-        return runScript;
+        let commandScript = `${portScript} . /mnt/frameworkbarrier/injector.sh && ${command}`;
+        return commandScript;
     }
     
     private async prepareRunScript(trialLocalTempFolder: string, curTrialSequenceId: number, trialJobId: string, trialWorkingFolder: string, form: JobApplicationForm): Promise<void> {
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
 
         await cpp.exec(`mkdir -p ${path.dirname(trialLocalTempFolder)}`);
-        await cpp.exec(`cp -r ${this.frameworkcontrollerTrialConfig.codeDir} ${trialLocalTempFolder}`);
+        await cpp.exec(`cp -r ${this.fcTrialConfig.codeDir} ${trialLocalTempFolder}`);
         const runScriptContent : string = CONTAINER_INSTALL_NNI_SHELL_FORMAT;
         // Write NNI installation file to local tmp files
         await fs.promises.writeFile(path.join(trialLocalTempFolder, 'install_nni.sh'), runScriptContent, { encoding: 'utf8' });
         // Create tmp trial working folder locally.
         await cpp.exec(`mkdir -p ${trialLocalTempFolder}`);
 
-        for(let taskRole of this.frameworkcontrollerTrialConfig.taskRoles) {
+        for(let taskRole of this.fcTrialConfig.taskRoles) {
             const runScriptContent: string = this.generateRunScript('frameworkcontroller', trialJobId, trialWorkingFolder, 
-                taskRole.command, curTrialSequenceId.toString(), taskRole.name, taskRole.gpuNum);
+            this.generateCommandScript(taskRole.command), curTrialSequenceId.toString(), taskRole.name, taskRole.gpuNum);
             await fs.promises.writeFile(path.join(trialLocalTempFolder, `run_${taskRole.name}.sh`), runScriptContent, { encoding: 'utf8' });
         }
 
@@ -232,12 +206,12 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
     
     private async prepareFrameworkControllerConfig(trialJobId: string, trialWorkingFolder: string, frameworkcontrollerJobName: string): Promise<any> {
 
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
 
         const podResources : any = [];
-        for(let taskRole of this.frameworkcontrollerTrialConfig.taskRoles) {
+        for(let taskRole of this.fcTrialConfig.taskRoles) {
             let resource: any = {};
             resource.requests = this.generatePodResource(taskRole.memoryMB, taskRole.cpuNum, taskRole.gpuNum);
             resource.limits = Object.assign({}, resource.requests);
@@ -280,14 +254,14 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
             case TrialConfigMetadataKey.TRIAL_CONFIG:
                 let frameworkcontrollerTrialJsonObjsect = JSON.parse(value);
 
-                this.frameworkcontrollerTrialConfig = new FrameworkControllerTrialConfig(
+                this.fcTrialConfig = new FrameworkControllerTrialConfig(
                     frameworkcontrollerTrialJsonObjsect.codeDir,
                     frameworkcontrollerTrialJsonObjsect.taskRoles
                 );
 
                 // Validate to make sure codeDir doesn't have too many files
                 try {
-                    await validateCodeDir(this.frameworkcontrollerTrialConfig.codeDir);
+                    await validateCodeDir(this.fcTrialConfig.codeDir);
                 } catch(error) {
                     this.log.error(error);
                     return Promise.reject(new Error(error));                    
@@ -301,13 +275,13 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
     }
     
     private generateContainerPort() {
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
 
         let port = 4000; //The default port used in container
-        for(let index in this.frameworkcontrollerTrialConfig.taskRoles) {
-            this.frameworkcontrollerContainerPortMap.set(this.frameworkcontrollerTrialConfig.taskRoles[index].name, port);
+        for(let index in this.fcTrialConfig.taskRoles) {
+            this.fcContainerPortMap.set(this.fcTrialConfig.taskRoles[index].name, port);
             port += 1;
         }
     }
@@ -324,29 +298,29 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
             throw new Error('frameworkcontroller Cluster config is not initialized');
         }
 
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
         
         let taskRoles = [];
-        for(let index in this.frameworkcontrollerTrialConfig.taskRoles) {
-            let containerPort = this.frameworkcontrollerContainerPortMap.get(this.frameworkcontrollerTrialConfig.taskRoles[index].name);
+        for(let index in this.fcTrialConfig.taskRoles) {
+            let containerPort = this.fcContainerPortMap.get(this.fcTrialConfig.taskRoles[index].name);
             if(!containerPort) {
                 throw new Error('Container port is not initialized');
             }
             let taskRole = this.generateTaskRoleConfig(
                 trialWorkingFolder, 
-                this.frameworkcontrollerTrialConfig.taskRoles[index].image, 
-                `run_${this.frameworkcontrollerTrialConfig.taskRoles[index].name}.sh`,
+                this.fcTrialConfig.taskRoles[index].image, 
+                `run_${this.fcTrialConfig.taskRoles[index].name}.sh`,
                 podResources[index],
                 containerPort
             );
             taskRoles.push({
-                name: this.frameworkcontrollerTrialConfig.taskRoles[index].name,
-                taskNumber: this.frameworkcontrollerTrialConfig.taskRoles[index].taskNum,
+                name: this.fcTrialConfig.taskRoles[index].name,
+                taskNumber: this.fcTrialConfig.taskRoles[index].taskNum,
                 frameworkAttemptCompletionPolicy: {
-                    minFailedTaskCount: this.frameworkcontrollerTrialConfig.taskRoles[index].frameworkAttemptCompletionPolicy.minFailedTaskCount, 
-                    minSucceededTaskCount: this.frameworkcontrollerTrialConfig.taskRoles[index].frameworkAttemptCompletionPolicy.minSucceededTaskCount
+                    minFailedTaskCount: this.fcTrialConfig.taskRoles[index].frameworkAttemptCompletionPolicy.minFailedTaskCount, 
+                    minSucceededTaskCount: this.fcTrialConfig.taskRoles[index].frameworkAttemptCompletionPolicy.minSucceededTaskCount
                 },
                 task: taskRole
             });
@@ -378,7 +352,7 @@ class FrameworkControllerTrainingService extends KubernetesTrainingService imple
             throw new Error('frameworkcontroller Cluster config is not initialized');
         }
 
-        if(!this.frameworkcontrollerTrialConfig) {
+        if(!this.fcTrialConfig) {
             throw new Error('frameworkcontroller trial config is not initialized');
         }
